+++ src/ft_put_ptr.c
@@ -6,7 +6,7 @@
 /*   By: dzhukov <dzhukov@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/01 22:23:55 by dzhukov           #+#    #+#             */
-/*   Updated: 2025/11/01 22:24:30 by dzhukov          ###   ########.fr       */
+/*   Updated: 2025/11/01 23:16:18 by dzhukov          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	ft_putptr(va_list *p_args)
 
 	count = write(1, "0x", 2);
 
-	ft_puthex(ptr, 'x');
+	ft_puthex((unsigned long long)ptr, 'x');
 
-	return (count + ft_hexlen(ptr));
+	return (count + ft_hexlen((unsigned long)ptr));
 }
